Upgrade Jest to v30.0.3 and resolve breaking changes (#378)
* Upgrade Jest to v30.0.3 and resolve breaking changes

- Update Jest from 29.6.4 to 30.0.3
- Update @types/jest from 29.5.4 to 30.0.0
- Update ts-jest from 29.1.1 to 29.4.0
- Update @types/node from 20.6.0 to 24.0.4
- Add support for new module file extensions (.mjs, .cjs, .mts, .cts)
- Update Jest configuration to include new test file patterns
- Fix Buffer type compatibility issue in WriteOutput.ts
- Update TypeScript lib to ES2022 for Array.at() support



* test-fix

* small-update

* workflow fix

---------

+++ packages/metaed-core/src/pipeline/WriteOutput.ts
@@ -22,7 +22,7 @@ function writeOutputFiles(result: GeneratorResult, outputDirectory: string) {
     if (output.resultString)
       fs.writeFileSync(`${outputDirectory}/${folderName}/${output.fileName}`, output.resultString, 'utf-8');
     else if (output.resultStream)
-      fs.writeFileSync(`${outputDirectory}/${folderName}/${output.fileName}`, output.resultStream);
+      fs.writeFileSync(`${outputDirectory}/${folderName}/${output.fileName}`, output.resultStream as Uint8Array);
     else Logger.debug(`No output stream or string for ${result.generatorName}`);
   });
 }
